Create tbl_file_modifier
this script is ment to modify the tbl file written by the annotation program prokka. When partial genes are annotated using prokka, the required 'partial gene' annotation doesn't make it to the tbl file. This script uses the .val file generated by tbl2asn to fix this partial gene annotation.

In addition, it fixes some common things in the annotation that cause errors in the submission process.
not a magic bullet, but it works for me.